requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
starlette==0.27.0
websockets==11.0.3

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.1.0.post1
bcrypt==4.1.2

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2

# Utilities
typing-extensions==4.9.0
watchfiles==0.21.0
httptools==0.6.1
uvloop==0.19.0

annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
greenlet==3.2.0
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pyasn1==0.4.8
pycparser==2.22
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
watchfiles==1.0.5
